Spun off generation to service
We donezo
